moved shared util code to outer util folder
