+++ cases-SEP24/transaction.test.js
@@ -1,12 +1,12 @@
 import { fetch } from "../util/fetchShim";
-import getSep10Token from "./util/sep10";
-import getTomlFile from "./util/getTomlFile";
+import getSep10Token from "../util/sep10";
+import getTomlFile from "../util/getTomlFile";
 import StellarSDK from "stellar-sdk";
-import { getTransactionBy } from "./util/transactions";
+import { getTransactionBy } from "../util/transactions";
 import { createTransaction } from "./util/interactive";
 import { errorSchema, getTransactionSchema } from "./util/schema";
 import { ensureCORS } from "../util/ensureCORS";
-import { getActiveCurrency } from "./util/currency";
+import { getActiveCurrency } from "../util/currency";
 
 jest.setTimeout(60000);
 
@@ -61,7 +61,7 @@ describe("Transaction", () => {
     });
     json = await getTransactionBy({
       value: json.id,
-      toml: toml,
+      transferServer: toml.TRANSFER_SERVER_SEP0024 || toml.TRANSFER_SERVER,
       jwt: null,
       expectStatusBetween: [400, 500],
     });
@@ -76,7 +76,11 @@ describe("Transaction", () => {
       jwt: jwt,
       isDeposit: true,
     });
-    json = await getTransactionBy({ value: json.id, toml: toml, jwt: jwt });
+    json = await getTransactionBy({
+      value: json.id,
+      transferServer: toml.TRANSFER_SERVER_SEP0024 || toml.TRANSFER_SERVER,
+      jwt: jwt,
+    });
     await checkTransactionResponse({ json: json, isDeposit: true });
   });
 
@@ -91,7 +95,7 @@ describe("Transaction", () => {
 
     json = await getTransactionBy({
       value: json.id,
-      toml: toml,
+      transferServer: toml.TRANSFER_SERVER_SEP0024 || toml.TRANSFER_SERVER,
       jwt: jwt,
     });
     await checkTransactionResponse({ json: json, isDeposit: false });
@@ -108,7 +112,7 @@ describe("Transaction", () => {
 
     json = await getTransactionBy({
       value: json.id,
-      toml: toml,
+      transferServer: toml.TRANSFER_SERVER_SEP0024 || toml.TRANSFER_SERVER,
       jwt: jwt,
     });
     const moreInfo = await fetch(json.transaction.more_info_url);
@@ -131,7 +135,7 @@ describe("Transaction", () => {
     const json = await getTransactionBy({
       value: "1277bd18-a2bd-4acd-9a87-2f541c7b8933",
       expectStatus: 404,
-      toml: toml,
+      transferServer: toml.TRANSFER_SERVER_SEP0024 || toml.TRANSFER_SERVER,
       jwt: jwt,
     });
     expect(json).toMatchSchema(errorSchema);
@@ -142,7 +146,7 @@ describe("Transaction", () => {
       iden: "stellar_transaction_id",
       value: "17a670bc424ff5ce3b386dbfaae9990b66a2a37b4fbe51547e8794962a3f9e6a",
       expectStatus: 404,
-      toml: toml,
+      transferServer: toml.TRANSFER_SERVER_SEP0024 || toml.TRANSFER_SERVER,
       jwt: jwt,
     });
     expect(json).toMatchSchema(errorSchema);
@@ -154,7 +158,7 @@ describe("Transaction", () => {
       value: "2dd16cb409513026fbe7defc0c6f826c2d2c65c3da993f747d09bf7dafd31093",
       expectStatus: 404,
       jwt: jwt,
-      toml: toml,
+      transferServer: toml.TRANSFER_SERVER_SEP0024 || toml.TRANSFER_SERVER,
     });
     expect(json).toMatchSchema(errorSchema);
   });
